feat(info-button): moving component to community #630
