+++ libs/angular-components/community/components/buttons/info-button/info-button.component.scss
