+++ libs/angular-components/community/components/buttons/info-button/info-button.stories.ts
@@ -1,8 +1,15 @@
 import { Meta, StoryObj } from "@storybook/angular";
 import { InfoButtonComponent } from "./info-button.component";
 
+/**
+ * <a href="https://www.figma.com/design/jWiRIXhHRxwVdMSimKX2FF/TEDI-READY-(work-in-progress)?node-id=4514-72997&m=dev" target="_BLANK">Figma ↗</a><br/>
+ * <a href="https://tedi.tehik.ee/1ee8444b7/p/0341c9-info-button" target="_BLANK">Zeroheight ↗</a>
+ * This is a simple info button component that can be used to display additional information when hovered over. It's mosty used together wihh tooltips or popovers.
+ * It can be used in various contexts, such as forms, dashboards, or any other UI where additional information is needed.
+ */
+
 export default {
-  title: "Tedi-Ready Angular/Components/Buttons/InfoButton",
+  title: "Community Angular/Buttons/InfoButton",
   impots: [],
   argTypes: {
     isSmall: {
