+++ libs/angular-components/community/public-api.ts
@@ -1,1 +1,2 @@
 export * from "./components/buttons/button/button.component";
+export * from "./components/buttons/info-button/info-button.component";
